feat(ui): split KVMSettings into composable components
