+++ ui/src/app/kvm/components/KVMConfigurationCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./KVMConfigurationCard";
