--- ui/src/app/kvm/components/KVMSettings/KVMConfiguration/KVMConfigurationFields/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./KVMConfigurationFields";
