--- ui/src/app/kvm/components/KVMSettings/KVMConfiguration/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./KVMConfiguration";
