--- ui/src/app/kvm/components/KVMSettings/KVMSettings.test.tsx
@@ -1,106 +1,0 @@
-import { mount } from "enzyme";
-import { Provider } from "react-redux";
-import { MemoryRouter } from "react-router-dom";
-import configureStore from "redux-mock-store";
-
-import KVMSettings from "./KVMSettings";
-
-import { KVMHeaderViews } from "app/kvm/constants";
-import { PodType } from "app/store/pod/constants";
-import type { RootState } from "app/store/root/types";
-import {
-  podDetails as podFactory,
-  podState as podStateFactory,
-  resourcePoolState as resourcePoolStateFactory,
-  rootState as rootStateFactory,
-  tagState as tagStateFactory,
-  zoneState as zoneStateFactory,
-} from "testing/factories";
-
-const mockStore = configureStore();
-
-describe("KVMSettings", () => {
-  let initialState: RootState;
-
-  beforeEach(() => {
-    initialState = rootStateFactory({
-      pod: podStateFactory({
-        items: [podFactory({ id: 1, name: "pod1" })],
-        loaded: true,
-      }),
-      resourcepool: resourcePoolStateFactory({
-        loaded: true,
-      }),
-      tag: tagStateFactory({
-        loaded: true,
-      }),
-      zone: zoneStateFactory({
-        loaded: true,
-      }),
-    });
-  });
-
-  it("fetches the necessary data on load", () => {
-    const state = { ...initialState };
-    const store = mockStore(state);
-    mount(
-      <Provider store={store}>
-        <MemoryRouter
-          initialEntries={[{ pathname: "/kvm/1/edit", key: "testKey" }]}
-        >
-          <KVMSettings id={1} setHeaderContent={jest.fn()} />
-        </MemoryRouter>
-      </Provider>
-    );
-    const expectedActions = [
-      "pod/fetch",
-      "resourcepool/fetch",
-      "tag/fetch",
-      "zone/fetch",
-    ];
-    const actions = store.getActions();
-    expectedActions.forEach((expectedAction) => {
-      expect(actions.some((action) => action.type === expectedAction));
-    });
-  });
-
-  it("displays a spinner if data has not loaded", () => {
-    const state = { ...initialState };
-    state.resourcepool.loaded = false;
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <MemoryRouter
-          initialEntries={[{ pathname: "/kvm/1/edit", key: "testKey" }]}
-        >
-          <KVMSettings id={1} setHeaderContent={jest.fn()} />
-        </MemoryRouter>
-      </Provider>
-    );
-    expect(wrapper.find("Spinner").length).toBe(1);
-  });
-
-  it("can open the delete form if the KVM is a LXD KVM", () => {
-    const state = { ...initialState };
-    state.pod.items[0].type = PodType.LXD;
-    const setHeaderContent = jest.fn();
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <MemoryRouter
-          initialEntries={[{ pathname: "/kvm/1/edit", key: "testKey" }]}
-        >
-          <KVMSettings id={1} setHeaderContent={setHeaderContent} />
-        </MemoryRouter>
-      </Provider>
-    );
-
-    wrapper.find("button[data-test='remove-kvm']").simulate("click");
-    expect(setHeaderContent).toHaveBeenCalledWith({
-      view: KVMHeaderViews.DELETE_KVM,
-      extras: {
-        hostId: 1,
-      },
-    });
-  });
-});
