--- ui/src/app/kvm/components/KVMSettings/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./KVMSettings";
