+++ ui/src/app/kvm/views/LXDSingleDetails/LXDSingleResources/LXDSingleResources.tsx
@@ -26,7 +26,7 @@ const LXDSingleResources = ({ id }: Props): JSX.Element => {
   }
   return (
     <>
-      <Strip className="u-no-padding--top" shallow>
+      <Strip shallow>
         <KVMResourcesCard id={pod.id} />
       </Strip>
       <Strip shallow>
