+++ ui/src/app/kvm/views/LXDSingleDetails/LXDSingleSettings/AuthenticationCard/UpdateCertificate/UpdateCertificateFields/UpdateCertificateFields.test.tsx
