+++ ui/src/app/kvm/views/LXDSingleDetails/LXDSingleSettings/DangerZoneCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DangerZoneCard";
