+++ ui/src/app/kvm/views/LXDSingleDetails/LXDSingleSettings/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LXDSingleSettings";
