+++ ui/src/app/kvm/views/VirshDetails/VirshDetails.tsx
@@ -6,10 +6,10 @@ import { Redirect, Route, Switch } from "react-router-dom";
 
 import VirshDetailsHeader from "./VirshDetailsHeader";
 import VirshResources from "./VirshResources";
+import VirshSettings from "./VirshSettings";
 
 import Section from "app/base/components/Section";
 import type { RouteParams } from "app/base/types";
-import KVMSettings from "app/kvm/components/KVMSettings";
 import { useActivePod } from "app/kvm/hooks";
 import type { KVMHeaderContent } from "app/kvm/types";
 import kvmURLs from "app/kvm/urls";
@@ -50,7 +50,7 @@ const VirshDetails = (): JSX.Element => {
             <VirshResources id={id} />
           </Route>
           <Route exact path={kvmURLs.virsh.details.edit(null, true)}>
-            <KVMSettings id={id} setHeaderContent={setHeaderContent} />
+            <VirshSettings id={id} />
           </Route>
           <Redirect
             from={kvmURLs.virsh.details.index(null, true)}
