+++ ui/src/app/kvm/views/VirshDetails/VirshSettings/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./VirshSettings";
